requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic-settings>=2.6.0
httpx>=0.27.0
openai>=1.55.0
streamlit>=1.40.0
pypdf>=5.0.0
python-docx>=1.1.0
pandas>=2.2.0
numpy>=2.0.0
weaviate-client>=4.9.0
